+++ packages/ui-components/src/components/MultiProgressBar/index.ts
@@ -1,2 +1,1 @@
 export * from "./MultiProgressBar";
-export * from "./interfaces";
